feat(documents): migrated documents controller
