+++ src/users/users.service.ts
@@ -5,23 +5,19 @@ import {
   Injectable,
   NotFoundException,
 } from '@nestjs/common';
+import { ConfigService } from '@nestjs/config/dist/config.service';
+import { hash } from 'bcrypt';
 import * as fs from 'fs';
 import * as path from 'path';
-import { User, USER_STATUS_TYPE } from 'src/models/user';
-import { getCurrentPeriod, getFileName } from 'src/utils/users.utils';
-import { OperationType, UpdateUserDTO } from './dto/updateData.dto';
-import { UserRegisterDTO } from './dto/userRegister.dto';
-import { createMatricula } from './utils';
-import { hash } from 'bcrypt';
-import { ConfigService } from '@nestjs/config/dist/config.service';
+import { ActivityHistoryService } from 'src/activity-history/activity-history.service';
+import { USER_STATUS_TYPE, User } from 'src/models/user';
 import {
   FileType,
-  FileTypeInterface,
-  UserDocuments,
+  UserDocuments
 } from 'src/models/user_documents';
-import { Documents_Observaciones } from 'src/models/documents_observaciones';
-import { ActivityHistoryService } from 'src/activity-history/activity-history.service';
-import { DOCUMENTS_OPREATIONS_MESSAGES } from 'src/contants';
+import { UpdateUserDTO } from './dto/updateData.dto';
+import { UserRegisterDTO } from './dto/userRegister.dto';
+import { createMatricula } from './utils';
 
 @Injectable()
 export class UsersService {
@@ -121,77 +117,9 @@ export class UsersService {
     return 'El usuario ha sido registrado con exito';
   }
 
-  async uploadDocument(
-    file: Express.Multer.File,
-    documentType: FileTypeInterface,
-    matricula: string,
-  ) {
-    try {
-      const user = await this.em.findOneOrFail(
-        User,
-        {
-          matricula,
-        },
-        { populate: ['documentos'] },
-      );
+ 
 
-      const document = user.documentos
-        .getItems()
-        .find((document) => document.fileType === documentType);
 
-      document.status = 'reviewing';
-
-      const newPath = path.join('./uploads', getCurrentPeriod(), matricula);
-
-      if (!fs.existsSync(newPath)) fs.mkdirSync(newPath, { recursive: true });
-
-      const FullPath = path.join(
-        newPath,
-        getFileName(documentType, file.originalname),
-      );
-
-      document.ruta = FullPath;
-
-      fs.writeFileSync(FullPath, file.buffer);
-
-      await this.em.persistAndFlush(document);
-
-      this.activityHistoryService.createActivityHistory({
-        action: 'update',
-        description: 'Se ha subido un nuevo documento',
-        updatedBy: user.id,
-        userAffected: user.id,
-      });
-
-      return {
-        message: 'El archivo se ha subido satisfactoriamente',
-        filename: file.filename,
-      };
-    } catch (error) {
-      console.error(error);
-      return {
-        message: error.message,
-      };
-    }
-  }
-
-  async findDocs(matricula: string) {
-    try {
-      const user = await this.em.findOneOrFail(
-        User,
-        {
-          matricula,
-        },
-        { populate: ['documentos', 'documentos.observaciones'] },
-      );
-
-      return {
-        documentos: user.documentos,
-      };
-    } catch (error) {
-      throw new NotFoundException('User not found');
-    }
-  }
 
   async createAdmin(userData: UserRegisterDTO, adminId: number) {
     try {
@@ -224,73 +152,6 @@ export class UsersService {
     }
   }
 
-  async findDocsById(id: number | string) {
-
-    const options = typeof id === 'number' ? { id } : { matricula: id };
-
-    try {
-      const user = await this.em.findOneOrFail(
-        User,
-        options,
-        { populate: ['documentos', 'documentos.observaciones'] },
-      );
-
-      return {
-        name: user.nombre,
-
-        documentos: user.documentos,
-      };
-    } catch (error) {
-      throw new NotFoundException('User not found');
-    }
-  }
-
-  async updateDocumentStatus(
-    id: number,
-    operation: OperationType,
-    adminId: number,
-    observaciones?: string[],
-  ) {
-    try {
-      const document = await this.em.findOneOrFail(
-        UserDocuments,
-        { id },
-        { populate: ['observaciones'] },
-      );
-
-      const user = await this.em.findOneOrFail(User, { documentos: { id } });
-
-      if (operation === 'approve') {
-        if (document.status === 'rejected') {
-          this.em.remove(document.observaciones.getItems());
-          document.observaciones.removeAll();
-          document.status = 'approved';
-        } else document.status = 'approved';
-      } else if (operation === 'reject') {
-        document.status = 'rejected';
-        document.observaciones.add(
-          observaciones.map((observacion) =>
-            this.em.create(Documents_Observaciones, { observacion }),
-          ),
-        );
-      }
-
-      await this.em.persistAndFlush(document);
-
-      this.activityHistoryService.createActivityHistory({
-        action: 'update',
-        description: DOCUMENTS_OPREATIONS_MESSAGES[operation],
-        updatedBy: adminId,
-        userAffected: user.id,
-      });
-
-      return {
-        message: 'El documento ha sido actualizado con exito',
-      };
-    } catch (error) {
-      throw new BadRequestException('No se pudo actualizar el documento');
-    }
-  }
 
   async findProspectos(status?: USER_STATUS_TYPE) {
     const statusOptions = status ? { status } : {};
